+++ src/main/java/com/example/trady/service/SellingService.java
@@ -9,7 +9,7 @@ import java.util.List;
 public interface SellingService {
     Selling createSelling(Selling selling);
 
-    Selling createSelling(Selling selling, Product product, String size, int price);
+    Selling createSelling(Selling selling, Product product, String size, long price);
 
     // 사용자별 판매 상품 조회 메서드
     List<Selling> findAllByUser(Member user);
